+++ mcpgen/__init__.py
@@ -1,3 +1,3 @@
 """MCPGen package."""
 
-__version__ = "1.0.0"
+__version__ = "1.1.0"